+++ packages/eslint-plugin-pf-codemods/test/rules/page-header-move-avatar.js
@@ -1,7 +1,7 @@
 const ruleTester = require("./ruletester");
-const rule = require("../../lib/rules/use-page-header-tools");
+const rule = require("../../lib/rules/page-header-move-avatar");
 
-ruleTester.run("use-page-header-tools", rule, {
+ruleTester.run("page-header-move-avatar", rule, {
   valid: [
     {
       code: `import { Avatar, PageHeader } from '@patternfly/react-core'; <PageHeader headerTools="tools" />`
@@ -21,6 +21,16 @@ ruleTester.run("use-page-header-tools", rule, {
           type: 'JSXElement'
         }
       ]
-    }
+    },
+    {
+      code: `import { Avatar, PageHeader, PageHeaderTools } from '@patternfly/react-core'; <PageHeader avatar={<Avatar />} headerTools={<PageHeaderTools></PageHeaderTools>} />`,
+      output: `import { Avatar, PageHeader, PageHeaderTools } from '@patternfly/react-core'; <PageHeader  headerTools={<PageHeaderTools><Avatar /></PageHeaderTools>} />`,
+      errors: [
+        {
+          message: 'avatar prop was removed from PageHeader, move the avatar component into PageHeaderTools',
+          type: 'JSXElement'
+        }
+      ]
+    },
   ]
 });
